move io thread handling into own module
